// src/lib/types.ts
export interface Testimonial {
    id: string;
    userId: string;
    userName: string;
    userAvatarColor: string;
    content: string;
    rating: number;
    createdAt: Timestamp;
    createdAtISO?: string;
    postedAt: string;
}